fix(fastcrc): use exact-width table reads
Problem:
The non-Arduino FastCRC pgm_read_dword helper loaded unsigned long values from uint32_t CRC tables. On LP64 that performed an eight-byte, eight-aligned access at each four-byte table entry, and a permissive sanitizer invocation could print the undefined access while still appearing green.

Evidence and decision:
At current base 18046ceac81caaf37f4be6fce436e2730a1edef9, the FastCRC blobs were unchanged from the issue evidence. The focused GCC LP64 reproducer stopped at FastCRCsw.cpp:48 with a misaligned unsigned long load and exit status 1. MicroStar-Tech/Livox-SDK2 commit 22d7d10e2918bfaa3e0b9a07532f19b41da1adbe was re-evaluated against this base. Its three exact-width hunks were adapted; sanitizer suppression, CRC seed changes, dependency replacement, and wholesale branch import were rejected.

Implementation:
Load word and dword table entries as uint16_t and uint32_t and declare the remaining unsigned-short CCITT table with uint16_t. Add a standalone CMake/CTest regression covering canonical profiles, optimized-loop boundaries, SDK header sizes, CRC32 payload lengths, pointer offsets, and a sanitizer negative control whose wrapper rejects stderr-only success.

Combined concerns:
The vendored FastCRC correction and focused tests span build/dependency/packaging and SDK/tests path categories, but are inseparable: the regression directly reproduces the same table access and certifies the behavior-preserving safety correction at real SDK lengths.

Compatibility:
Public API, ABI, wire layout, packet format, packaging, and intended consumer results are unchanged. The canonical values remain CCITT-FALSE 0x29B1, Livox mcrf4xx 0x2189, and CRC-32/ISO-HDLC 0xCBF43926; standard MCRF4XX 0x6F91 remains a separately named reference. The exact-width helpers are defined consistently for ILP32, LP64, and LLP64 data models.

Verification:
The old GCC LP64 ASan/UBSan build reproduced the line 48 diagnostic and exited 1. The corrected standalone GCC 15.2 CMake build passed 2/2 sanitizer CTests, including the fail-closed negative control, and 1/1 Release CTest without sanitizers. The complete SDK Release build reached 100%, governance unit tests passed 35/35, and git diff --check passed. AppleClang and MSVC execution remain pending. An ILP32 -m32 attempt remains pending because this environment lacks multilib libc headers.

Source attribution:
The exact-width source adaptation derives from MicroStar-Tech/Livox-SDK2 commit 22d7d10e2918bfaa3e0b9a07532f19b41da1adbe, authored by Xuqing Liu <ciwei100000@gmail.com>, under the compatible existing FastCRC MIT license. Frank Bösing source notices remain intact. The focused regression is downstream project work authored by OpenAI Codex.

Upstream disposition:
Apply the safety fix downstream without waiting for upstream. Original-FastCRC submissions for the direct size_t include and exact-width table helpers are deferred to maintainer pOmelchenko after the draft PR platform matrix and final review complete; those objective triggers allow a focused upstream patch without blocking downstream qualification.

Refs: #6
Agent-Authored: OpenAI Codex

+++ 3rdparty/FastCRC/FastCRCsw.cpp
@@ -39,13 +39,16 @@
 #if !defined(ARDUINO)
 #define PROGMEM
 #define pgm_read_byte(addr) (*(const unsigned char *)(addr))
+
+#include <stdint.h>
+
 template <typename T>
-static constexpr unsigned short pgm_read_word(T addr) noexcept {
-	return *(const unsigned short *)(addr);
+static constexpr uint16_t pgm_read_word(T addr) noexcept {
+	return *(const uint16_t *)(addr);
 };
 template <typename T>
-static constexpr unsigned long pgm_read_dword(T addr) noexcept {
-	return *(const unsigned long *)(addr);
+static constexpr uint32_t pgm_read_dword(T addr) noexcept {
+	return *(const uint32_t *)(addr);
 };
 #endif
